requirements.txt
abstract_singleton==1.0.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
apache-beam==2.38.0
apache-flink==1.16.0
apache-flink-libraries==1.16.0
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
autopep8==2.0.1
avro-python3==1.9.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
blinker==1.6.2
blis==0.7.9
boto3==1.18.49
botocore==1.21.65
CacheControl==0.12.11
cachetools==4.2.4
cachy==0.3.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.0.12
cleo==1.0.0a5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.1.0
colorama==0.4.6
confection==0.0.4
confluent-kafka==1.9.2
confuse==1.4.0
contourpy==1.0.7
coverage==7.2.3
crashtest==0.3.1
crcmod==1.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
datadog==0.44.0
datadog-api-client==2.6.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.6
distro==1.8.0
Django==4.1.3
django-environ==0.9.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-api-key==1.0.0
dnspython==2.3.0
docker==6.0.1
docopt==0.6.2
duckduckgo-search==2.8.6
dulwich==0.20.50
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.1.1
executing==1.2.0
ez-setup==0.9
Faker==17.0.0
fastavro==1.4.7
filelock==3.8.0
find-libpython==0.3.0
flake8==6.0.0
Flask==2.3.2
fonttools==4.39.0
frozenlist==1.3.3
funcy==1.16
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
gTTS==2.3.1
h11==0.14.0
hdfs==2.7.0
html5lib==1.1
httpcore==0.17.0
httplib2==0.19.1
httpx==0.24.0
identify==2.5.22
idna==3.4
ijson==3.2.0.post0
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.7.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.0
keyring==23.11.0
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
lockfile==0.12.2
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.2
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nodeenv==1.7.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
orjson==3.8.4
outcome==1.2.0
packaging==23.1
pandas==1.3.5
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pemja==0.2.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pipenv==2022.11.11
pipenv-poetry-migrate==0.2.1
pkginfo==1.8.3
platformdirs==2.5.4
playsound==1.2.2
pluggy==1.0.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
pre-commit==3.2.2
preshed==3.0.8
prompt-toolkit==3.0.32
proto-plus==1.19.6
protobuf==4.22.3
psycopg==3.1.8
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
pyflakes==3.0.1
Pygments==2.13.0
pylev==1.4.0
pymongo==3.13.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-vcr==1.0.2
python-amazon-sp-api==0.18.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0
pyyaml_env_tag==0.1
readability-lxml==0.8.1
redis==3.5.3
regex==2023.3.23
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.5.2
selenium==4.1.4
shellingham==1.5.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spotipy==2.23.0
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
thinc==8.1.9
tiktoken==0.3.3
timeseries==0.5.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
urllib3-secure-extra==0.1.0
vcrpy==4.2.1
vine==5.0.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.5
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
wrapt==1.14.1
wsproto==1.2.0
xattr==0.9.9
yarl==1.9.1
zipp==3.10.0
